Refatoração da estrutura do projeto
Routers criados para Users e Auth; constantes movidas para variáveis de ambiente
