+++ src/fast_zero/app.py
@@ -1,145 +1,20 @@
 from http import HTTPStatus
 
-from fastapi import Depends, FastAPI, HTTPException
+from fastapi import FastAPI
 from fastapi.responses import HTMLResponse
-from fastapi.security import OAuth2PasswordRequestForm
-from sqlalchemy import select
-from sqlalchemy.orm import Session
 
-from fast_zero.database import get_session
-from fast_zero.models import User
-from fast_zero.schemas import Message, Token, UserList, UserPublic, UserSchema
-from fast_zero.security import (
-    create_access_token,
-    get_current_user,
-    get_password_hash,
-    verify_password,
-)
+from fast_zero.routers import auth, users
+from fast_zero.schemas import Message
 
 app = FastAPI()
+
+app.include_router(auth.router)
+app.include_router(users.router)
 
 
 @app.get('/', status_code=HTTPStatus.OK, response_model=Message)
 def read_root() -> dict:
     return {'message': 'Olá, mundo!!'}
-
-
-@app.post('/users/', status_code=HTTPStatus.CREATED, response_model=UserPublic)
-def create_user(
-    user: UserSchema, session: Session = Depends(get_session)
-) -> UserPublic:
-    db_user = session.scalar(
-        select(User).where(
-            (User.username == user.username) | (User.email == user.email)
-        )
-    )
-
-    if db_user:
-        if db_user.username == user.username:
-            raise HTTPException(
-                status_code=HTTPStatus.BAD_REQUEST,
-                detail='Username already exists',
-            )
-        if db_user.email == user.email:
-            raise HTTPException(
-                status_code=HTTPStatus.BAD_REQUEST,
-                detail='Email already exists',
-            )
-
-    db_user = User(
-        username=user.username,
-        email=user.email,
-        password=get_password_hash(user.password),
-    )
-
-    session.add(db_user)
-    session.commit()
-    session.refresh(db_user)
-
-    return db_user
-
-
-@app.get('/users/', response_model=UserList)
-def read_users(
-    limit: int = 10, offset: int = 0, session: Session = Depends(get_session)
-) -> dict:
-    users = session.scalars(select(User).limit(limit).offset(offset))
-
-    return {'users': users}
-
-
-@app.get('/users/{user_id}', response_model=UserPublic)
-def read_user_by_id(
-    user_id: int, session: Session = Depends(get_session)
-) -> dict:
-    user_by_id = session.get(User, user_id)
-
-    if not user_by_id:
-        raise HTTPException(
-            status_code=HTTPStatus.NOT_FOUND, detail='User not found'
-        )
-
-    return user_by_id
-
-
-@app.put('/users/{user_id}', response_model=UserPublic)
-def update_users(
-    user_id: int,
-    user: UserSchema,
-    session: Session = Depends(get_session),
-    current_user=Depends(get_current_user),
-) -> UserPublic:
-    if current_user.id != user_id:
-        raise HTTPException(
-            status_code=HTTPStatus.FORBIDDEN, detail='Not enough permission'
-        )
-
-    current_user.username = user.username
-    current_user.email = user.email
-    current_user.password = get_password_hash(user.password)
-
-    session.add(current_user)
-    session.commit()
-    session.refresh(current_user)
-
-    return current_user
-
-
-@app.delete('/users/{user_id}', response_model=Message)
-def delete_user(
-    user_id: int,
-    session: Session = Depends(get_session),
-    current_user=Depends(get_current_user),
-) -> dict:
-    if current_user.id != user_id:
-        raise HTTPException(
-            status_code=HTTPStatus.FORBIDDEN, detail='Not enough permission'
-        )
-
-    session.delete(current_user)
-    session.commit()
-
-    return {'message': 'User deleted'}
-
-
-@app.post('/token', response_model=Token)
-def login_for_acess_token(
-    form_data: OAuth2PasswordRequestForm = Depends(),
-    session: Session = Depends(get_session),
-) -> dict:
-    user = session.scalar(
-        select(User).where(User.username == form_data.username)
-    )
-
-    if not user or not verify_password(form_data.password, user.password):
-        raise HTTPException(
-            status_code=HTTPStatus.BAD_REQUEST,
-            detail='Incorrect username or password',
-        )
-
-    access_token = create_access_token(data={'sub': user.username})
-
-    return {'access_token': access_token, 'token_type': 'Bearer'}
 
 
 @app.get('/pagina', status_code=HTTPStatus.OK, response_class=HTMLResponse)
