+++ src/fast_zero/routers/auth.py
@@ -1,0 +1,36 @@
+from http import HTTPStatus
+from typing import Annotated
+
+from fastapi import APIRouter, Depends, HTTPException
+from fastapi.security import OAuth2PasswordRequestForm
+from sqlalchemy import select
+from sqlalchemy.orm import Session
+
+from fast_zero.database import get_session
+from fast_zero.models import User
+from fast_zero.schemas import Token
+from fast_zero.security import create_access_token, verify_password
+
+router = APIRouter(prefix='/auth', tags=['auth'])
+OAuth2Form = Annotated[OAuth2PasswordRequestForm, Depends()]
+Session = Annotated[Session, Depends(get_session)]
+
+
+@router.post('/token', response_model=Token)
+def login_for_acess_token(
+    form_data: OAuth2Form,
+    session: Session,  # type: ignore
+) -> dict:
+    user = session.scalar(
+        select(User).where(User.username == form_data.username)
+    )
+
+    if not user or not verify_password(form_data.password, user.password):
+        raise HTTPException(
+            status_code=HTTPStatus.BAD_REQUEST,
+            detail='Incorrect username or password',
+        )
+
+    access_token = create_access_token(data={'sub': user.username})
+
+    return {'access_token': access_token, 'token_type': 'Bearer'}
