+++ src/fast_zero/routers/users.py
@@ -1,0 +1,119 @@
+from http import HTTPStatus
+from typing import Annotated
+
+from fastapi import APIRouter, Depends, HTTPException, Query
+from sqlalchemy import select
+from sqlalchemy.orm import Session
+
+from fast_zero.database import get_session
+from fast_zero.models import User
+from fast_zero.schemas import (
+    FilterPage,
+    Message,
+    UserList,
+    UserPublic,
+    UserSchema,
+)
+from fast_zero.security import get_current_user, get_password_hash
+
+router = APIRouter(prefix='/users', tags=['users'])
+Session = Annotated[Session, Depends(get_session)]
+CurrentUser = Annotated[User, Depends(get_current_user)]
+FilterUser = Annotated[FilterPage, Query()]
+
+
+@router.post('/', status_code=HTTPStatus.CREATED)
+def create_user(user: UserSchema, session: Session) -> UserPublic:  # type: ignore
+    db_user = session.scalar(
+        select(User).where(
+            (User.username == user.username) | (User.email == user.email)
+        )
+    )
+
+    if db_user:
+        if db_user.username == user.username:
+            raise HTTPException(
+                status_code=HTTPStatus.BAD_REQUEST,
+                detail='Username already exists',
+            )
+        if db_user.email == user.email:
+            raise HTTPException(
+                status_code=HTTPStatus.BAD_REQUEST,
+                detail='Email already exists',
+            )
+
+    db_user = User(
+        username=user.username,
+        email=user.email,
+        password=get_password_hash(user.password),
+    )
+
+    session.add(db_user)
+    session.commit()
+    session.refresh(db_user)
+
+    return db_user
+
+
+@router.get('/', response_model=UserList)
+def read_users(
+    session: Session,  # type: ignore
+    filter_users: FilterUser,
+):
+    users = session.scalars(
+        select(User).limit(filter_users.limit).offset(filter_users.offset)
+    )
+
+    return {'users': users}
+
+
+@router.get('/{user_id}')
+def read_user_by_id(user_id: int, session: Session) -> UserPublic:  # type: ignore
+    user_by_id = session.get(User, user_id)
+
+    if not user_by_id:
+        raise HTTPException(
+            status_code=HTTPStatus.NOT_FOUND, detail='User not found'
+        )
+
+    return user_by_id
+
+
+@router.put('/{user_id}')
+def update_users(
+    user_id: int,
+    user: UserSchema,
+    session: Session,  # type: ignore
+    current_user: CurrentUser,
+) -> UserPublic:
+    if current_user.id != user_id:
+        raise HTTPException(
+            status_code=HTTPStatus.FORBIDDEN, detail='Not enough permission'
+        )
+
+    current_user.username = user.username
+    current_user.email = user.email
+    current_user.password = get_password_hash(user.password)
+
+    session.add(current_user)
+    session.commit()
+    session.refresh(current_user)
+
+    return current_user
+
+
+@router.delete('/{user_id}', response_model=Message)
+def delete_user(
+    user_id: int,
+    session: Session,  # type: ignore
+    current_user: CurrentUser,
+) -> dict:
+    if current_user.id != user_id:
+        raise HTTPException(
+            status_code=HTTPStatus.FORBIDDEN, detail='Not enough permission'
+        )
+
+    session.delete(current_user)
+    session.commit()
+
+    return {'message': 'User deleted'}
