+++ src/fast_zero/schemas.py
@@ -25,3 +25,8 @@ class UserPublic(BaseModel):
 
 class UserList(BaseModel):
     users: list[UserPublic]
+
+
+class FilterPage(BaseModel):
+    offset: int = 0
+    limit: int = 100
