+++ tests/test_security.py
@@ -2,14 +2,17 @@ from http import HTTPStatus
 
 from jwt import decode
 
-from fast_zero.security import ALGORITHM, SECRET_KEY, create_access_token
+from fast_zero.security import create_access_token
+from fast_zero.settings import Settings
+
+settings = Settings()
 
 
 def test_jwt():
     data = {'sub': 'torresminho'}
     token = create_access_token(data)
 
-    result = decode(token, SECRET_KEY, algorithms=ALGORITHM)
+    result = decode(token, settings.SECRET_KEY, algorithms=settings.ALGORITHM)
 
     assert result['sub'] == data['sub']
     assert result['exp']
